fixes: small changes to profile page

+++ app/routes/seed.jsx
@@ -1,21 +1,16 @@
 import { json } from "@remix-run/node";
-import { Form, useActionData, useLoaderData } from "@remix-run/react";
+import { Form, Link, useActionData, useLoaderData } from "@remix-run/react";
 import connectDb from "~/db/connectDb.server";
 import Navbar from "~/components/Navbar";
 import { getUserId } from "../utils/auth.server";
 
 export const loader = async ({ request }) => {
   const userId = await getUserId(request);
-
-  if (!userId) {
-    return json({ error: "Not logged in" }, { status: 401 });
-  }
-
   const db = await connectDb();
   const snippets = await db.models.snippets.countDocuments();
   const snippetFolders = await db.models.snippetFolders.countDocuments();
 
-  return json({ snippets, snippetFolders });
+  return json({ snippets, snippetFolders, userId });
 };
 
 export default function Seed() {
@@ -28,7 +23,7 @@ export default function Seed() {
       <div className="mt-24 text-center">
         <h1 className="text-2xl font-bold">Seed</h1>
         <div className="bg-gray-200 p-4 rounded mt-4">
-          {loaderData.error && <p>{loaderData.error}</p>}
+          {!loaderData.userId && <p className="text-red-500">You must be logged in to seed the database.</p>}
           <p>
             Database has <span className="text-red-500 font-bold">{loaderData.snippets}</span> snippets and{" "}
             <span className="text-red-500 font-bold">{loaderData.snippetFolders}</span> snippet folders.
@@ -39,8 +34,25 @@ export default function Seed() {
               Seed
             </button>
           </Form>
-          {actionData?.success && <p className="text-green-500">{actionData.success}</p>}
-          {actionData?.error && <p className="text-red-500">{actionData.error}</p>}
+
+          {actionData?.success && (
+            <>
+              <p className="text-green-500">{actionData.success}</p>
+              <br/>
+              <Link to="/snippets" className="mt-4 bg-blue-500 text-white px-4 py-2 rounded">
+                View snippets
+              </Link>
+            </>
+          )}
+          {actionData?.error && (
+            <>
+              <p className="text-red-500">{actionData.error}</p>
+              <br/>
+              <Link to="/login" className="mt-4 bg-blue-500 text-white px-4 py-2 rounded">
+                Login
+              </Link>
+            </>
+          )}
         </div>
       </div>
     </div>
@@ -50,10 +62,13 @@ export default function Seed() {
 export async function action({ request }) {
   // Seed the database with some data.
   try {
+    const userId = await getUserId(request);
+    if (!userId) {
+      return json({ error: "You must be logged in to seed the database." }, { status: 401 });
+    }
     const db = await connectDb();
     await db.models.snippets.deleteMany({});
     await db.models.snippetFolders.deleteMany({});
-    const userId = await getUserId(request);
 
     const folder1 = await db.models.snippetFolders.create({
       name: "My Snippets",
